requirements.txt
# File: requirements.txt
streamlit
google-genai
httpx